Updated Modal Styles
