+++ client/src/containers/User/Dashboard/Dashboard.css
@@ -11,19 +11,3 @@
 .Profile h1 {
   font-size: 1.4em;
   padding-left: 10px; }
-
-.Modal {
-  position: absolute;
-  top: 120px;
-  left: 120px;
-  right: 120px;
-  bottom: 40px;
-  background-color: white; }
-
-.Overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rebeccapurple; }
