+++ client/src/containers/User/Dashboard/Dashboard.scss
@@ -14,22 +14,3 @@
   font-size: 1.4em;
   padding-left: 10px;
 }
-
-//  Modal
-.Modal {
-  position: absolute;
-  top: 120px;
-  left: 120px;
-  right: 120px;
-  bottom: 40px;
-  background-color: white;
-}
-
-.Overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rebeccapurple;
-}
