+++ client/src/hoc/ModalHOC/ModalHOC.scss
@@ -1,0 +1,19 @@
+//  Modal
+.Modal {
+  position: absolute;
+  
+  top: 120px;
+  left: 40px;
+  right: 40px;
+  bottom: 40px;
+  background-color: white;
+}
+
+.Overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: rgba(0, 0, 0, 0.5);
+}
